Migrated core PropertyTypeResolver logic to service impl
