+++ src/test/java/org/openmrs/plugin/rest/analyzer/introspection/SchemaIntrospectionService.java
@@ -12,6 +12,8 @@ package org.openmrs.plugin.rest.analyzer.introspection;
 import java.util.Map;
 
 import org.openmrs.module.webservices.rest.web.resource.api.Resource;
+import org.openmrs.module.webservices.rest.web.resource.impl.DelegatingResourceDescription;
+import org.openmrs.module.webservices.rest.web.resource.impl.DelegatingResourceHandler;
 
 /**
  * Service for introspecting resource properties via reflection.
@@ -43,4 +45,23 @@ public interface SchemaIntrospectionService {
 	 * @return A map of property names to their Java type names
 	 */
 	Map<String, String> discoverResourceProperties(Resource resource);
+	
+	/**
+	 * Determines the most accurate property type using multiple resolution strategies:
+	 * 1. Introspected types from delegate classes (highest priority)
+	 * 2. Representation metadata from DelegatingResourceDescription.Property
+	 * 3. Reflection on delegate class properties
+	 * 4. Intelligent inference from property names
+	 * 5. Fallback to generic types
+	 * 
+	 * @param propertyName The name of the property
+	 * @param property The DelegatingResourceDescription.Property containing representation metadata
+	 * @param handler The resource handler
+	 * @param introspectedProperties Map of properties discovered via reflection on delegate class
+	 * @return The most accurate type string for this property
+	 */
+	String determineAccuratePropertyType(String propertyName, 
+	                                   DelegatingResourceDescription.Property property,
+	                                   DelegatingResourceHandler<?> handler,
+	                                   Map<String, String> introspectedProperties);
 }
